# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ME Piston Component Running Hours Record API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vessels/{vesselId}/movements:
    get:
      operationId: listMovements
      tags: [movements]
      summary: List movement log entries (last 200)
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movement log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovementEntry"
    post:
      operationId: recordMovement
      tags: [movements]
      summary: Record a component movement
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovementInput"
      responses:
        "201":
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovementEntry"
        "400":
          description: Validation error
  /vessels/{vesselId}/movements/{movementId}:
    delete:
      operationId: deleteMovement
      tags: [movements]
      summary: Delete a movement log entry
      parameters:
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
        - name: movementId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Dashboard & Alerts ────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{userId}/vessel-access:
    get:
      operationId: listUserVesselAccess
      tags: [users]
      summary: List the vessels a user has explicit access to
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vessel access list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserVesselAccessItem"
    post:
      operationId: grantVesselAccess
      tags: [users]
      summary: Grant a user access to a vessel
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantVesselAccessInput"
      responses:
        "201":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserVesselAccessItem"
  /users/{userId}/vessel-access/{vesselId}:
    delete:
      operationId: revokeVesselAccess
      tags: [users]
      summary: Revoke a user's access to a vessel
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: vesselId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Access revoked
components:
  schemas:
...
    MovementEntry:
      type: object
      required: [id, vesselId, movementDate, meRh, componentId, fromLocation, toLocation, action, rhAdded, createdAt]
      properties:
        id:
          type: integer
        vesselId:
          type: integer
        movementDate:
          type: string
        meRh:
          type: number
        componentId:
          type: string
        fromLocation:
          type: string
        toLocation:
          type: string
        action:
          type: string
        rhAdded:
          type: number
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    MovementInput:
      type: object
      required: [movementDate, meRh, componentId, fromLocation, toLocation, action]
      properties:
        movementDate:
          type: string
        meRh:
          type: number
          minimum: 0
        componentId:
          type: string
        fromLocation:
          type: string
        toLocation:
          type: string
        action:
          type: string
        remarks:
          type: string

    # ─── Dashboard ───────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, email, fullName, role, isActive]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [vessel_officer, technical_office]
        isActive:
          type: boolean
        vesselIds:
          type: array
          items:
            type: integer
          description: "Accessible vessel ids; omitted/empty for technical_office (implicit fleet-wide access)"
    CreateUserInput:
      type: object
      required: [email, password, fullName, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        role:
          type: string
          enum: [vessel_officer, technical_office]
    UpdateUserInput:
      type: object
      properties:
        fullName:
          type: string
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [vessel_officer, technical_office]
        isActive:
          type: boolean
    UserVesselAccessItem:
      type: object
      required: [vesselId, vesselName]
      properties:
        vesselId:
          type: integer
        vesselName:
          type: string
    GrantVesselAccessInput:
      type: object
      required: [vesselId]
      properties:
        vesselId:
          type: integer